mcp/requirements.txt
mcp[cli]>=1.3.0
groq>=0.9.0
duckdb>=1.1.0
boto3>=1.34.0
python-dotenv>=1.0.0
uvicorn>=0.30.0
starlette>=0.40.0
fastapi>=0.111.0
pandas>=2.2.0
